requirements.txt
alembic==1.7.5
click==8.0.3
Flask==2.2.2
Flask-Migrate==3.1.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-SQLAlchemy==2.5.1
greenlet==2.0.0a1
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.1.1
psycopg2==2.9.3
psycopg2-binary
SQLAlchemy==1.4.31
Werkzeug==2.2.2
marshmallow==3.14.1
marshmallow-sqlalchemy==0.27.0
python-dotenv==1.0.0
gunicorn==20.1.0